Added navigation (without style) and moved Voertuig/Tankkaart/bestuurder in their own folders.
