+++ src/app/app.home.html
@@ -1,7 +1,5 @@
 <div>
-  <h1>Here comes the header and navigation</h1>
+  <app-navigation></app-navigation>
+  <router-outlet></router-outlet>
 </div>
 
-<div>
-  <app-voertuig></app-voertuig>
-</div>
